Move git code into test-runner
