+++ src/tester/get-affected-packages.ts
@@ -1,8 +1,5 @@
-import { parseMajorVersionFromDirectoryName } from "../lib/definition-parser";
-import { AllPackages, PackageBase, TypingsData, PackageId, DependencyVersion, getMangledNameForScopedPackage } from "../lib/packages";
-import { sourceBranch, typesDirectoryName } from "../lib/settings";
-import { Logger } from "../util/logging";
-import { execAndThrowErrors, flatMap, mapDefined, mapIter, sort } from "../util/util";
+import { AllPackages, PackageBase, TypingsData, PackageId, getMangledNameForScopedPackage } from "../lib/packages";
+import { mapDefined, mapIter, sort } from "../util/util";
 
 export interface GitDiff {
     status: "A" | "D" | "M";
@@ -98,92 +95,4 @@ function getReverseDependencies(allPackages: AllPackages, changedPackages: Packa
 
 function packageIdToKey(pkg: PackageId): string {
     return getMangledNameForScopedPackage(pkg.name) + "/v" + pkg.majorVersion;
-}
-
-/** Returns all immediate subdirectories of the root directory that have changed. */
-export function gitChanges(diffs: GitDiff[]): PackageId[] {
-    const changedPackages = new Map<string, Set<DependencyVersion>>();
-
-    for (const diff of diffs) {
-        const dep = getDependencyFromFile(diff.file);
-        if (dep) {
-            const versions = changedPackages.get(dep.name);
-            if (!versions) {
-                changedPackages.set(dep.name, new Set([dep.majorVersion]));
-            } else {
-                versions.add(dep.majorVersion);
-            }
-        }
-    }
-
-    return Array.from(flatMap(changedPackages, ([name, versions]) =>
-        mapIter(versions, majorVersion => ({ name, majorVersion }))));
-}
-
-/*
-We have to be careful about how we get the diff because travis uses a shallow clone.
-
-Travis runs:
-    git clone --depth=50 https://github.com/DefinitelyTyped/DefinitelyTyped.git DefinitelyTyped
-    cd DefinitelyTyped
-    git fetch origin +refs/pull/123/merge
-    git checkout -qf FETCH_HEAD
-
-If editing this code, be sure to test on both full and shallow clones.
-*/
-export async function gitDiff(log: Logger, definitelyTypedPath: string): Promise<GitDiff[]> {
-    try {
-        await run(`git rev-parse --verify ${sourceBranch}`);
-        // If this succeeds, we got the full clone.
-    } catch (_) {
-        // This is a shallow clone.
-        await run(`git fetch origin ${sourceBranch}`);
-        await run(`git branch ${sourceBranch} FETCH_HEAD`);
-    }
-
-    let diff = (await run(`git diff ${sourceBranch} --name-status`)).trim();
-    if (diff === "") {
-        // We are probably already on master, so compare to the last commit.
-        diff = (await run(`git diff ${sourceBranch}~1 --name-status`)).trim();
-    }
-    return diff.split("\n").map(line => {
-        var [status, file] = line.split(/\s+/, 2);
-        return { status: status.trim(), file: file.trim() } as GitDiff;
-    });
-
-    async function run(cmd: string): Promise<string> {
-        log(`Running: ${cmd}`);
-        const stdout = await execAndThrowErrors(cmd, definitelyTypedPath);
-        log(stdout);
-        return stdout;
-    }
-}
-
-/**
- * For "types/a/b/c", returns { name: "a", version: "*" }.
- * For "types/a/v3/c", returns { name: "a", version: 3 }.
- * For "x", returns undefined.
- */
-function getDependencyFromFile(file: string): PackageId | undefined {
-    const parts = file.split("/");
-    if (parts.length <= 2) {
-        // It's not in a typings directory at all.
-        return undefined;
-    }
-
-    const [typesDirName, name, subDirName] = parts; // Ignore any other parts
-
-    if (typesDirName !== typesDirectoryName) {
-        return undefined;
-    }
-
-    if (subDirName) {
-        // Looks like "types/a/v3/c"
-        const majorVersion = parseMajorVersionFromDirectoryName(subDirName);
-        if (majorVersion !== undefined) {
-            return { name,  majorVersion };
-        }
-    }
-
-    return { name, majorVersion: "*" };
 }
